+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - New sub-command: `/m cmd [arguments]` to fix command conflicts with server-side commands with the same name `/m`
   - e.g. `/m cmd hello world` executes the server command `/m hello world`
 - Added `/<command with tab-completable username> say [optional text]`: You can e.g. say `f` again without triggering the server-side command `/f` by typing `/f say`
+- Dungeon party finder: Added a warning message if you enter a floor other than the one you have queued for 
 
 ## [1.8.9-0.10.1] - 06.08.2020
 ### Added